move action manager
