+++ napari/_qt/qt_main_window.py
@@ -27,8 +27,8 @@ from qtpy.QtWidgets import (
 from .. import plugins
 from ..utils import config, perf
 from ..utils.history import get_save_history, update_save_history
+from ..utils.action_manager import action_manager
 from ..utils.io import imsave
-from ..utils.key_bindings import action_manager
 from ..utils.misc import in_jupyter, running_as_bundled_app
 from ..utils.settings import SETTINGS
 from ..utils.translations import trans
